chatbot/requirements.txt
streamlit>=1.22.0
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-docx>=0.8.11
langchain>=0.0.238
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
python-dotenv>=1.0.0
